Juste pour faire en sorte de pas avoir Version 0.42 a chaque update si cela n'est pas necessaire
 Modified Files:
 	glpi/update.php

+++ update.php
@@ -822,7 +822,20 @@ if(!FieldExists("glpi_config","permit_helpdesk")) {
 	$db->query($query) or die("glpi_config_permit_helpdesk ".$lang["update"][90].$db->error());
 }
 
+if(!TableExists("glpi_reservation_item")) {
+
  echo "<br>Version 0.42 <br />";
+
+	$query = "CREATE TABLE glpi_reservation_item (ID int(11) NOT NULL auto_increment,device_type tinyint(4) NOT NULL default '0', id_device int(11) NOT NULL default '0', comments text NOT NULL, PRIMARY KEY  (ID), KEY device_type (device_type));";
+
+	$db->query($query) or die("4201 ".$lang["update"][90].$db->error());
+}
+
+if(!TableExists("glpi_reservation_resa")) {
+	$query = "CREATE TABLE glpi_reservation_resa (  ID bigint(20) NOT NULL auto_increment,  id_item int(11) NOT NULL default '0',  begin datetime NOT NULL default '0000-00-00 00:00:00',  end datetime NOT NULL default '0000-00-00 00:00:00',  id_user int(11) NOT NULL default '0',  PRIMARY KEY  (ID),  KEY id_item (id_item),  KEY id_user (id_user),  KEY begin (begin),  KEY end (end));";
+
+	$db->query($query) or die("4202 ".$lang["update"][90].$db->error());
+}
 
 //Mise a jour 0.42 verification des prefs pour chaque user.
 $query = "select ID, name from glpi_users";
@@ -844,18 +857,6 @@ if($db->numrows($result) != $db->numrows($result2)) {
 	}
 }
 
-if(!TableExists("glpi_reservation_item")) {
-
-	$query = "CREATE TABLE glpi_reservation_item (ID int(11) NOT NULL auto_increment,device_type tinyint(4) NOT NULL default '0', id_device int(11) NOT NULL default '0', comments text NOT NULL, PRIMARY KEY  (ID), KEY device_type (device_type));";
-
-	$db->query($query) or die("4201 ".$lang["update"][90].$db->error());
-}
-
-if(!TableExists("glpi_reservation_resa")) {
-	$query = "CREATE TABLE glpi_reservation_resa (  ID bigint(20) NOT NULL auto_increment,  id_item int(11) NOT NULL default '0',  begin datetime NOT NULL default '0000-00-00 00:00:00',  end datetime NOT NULL default '0000-00-00 00:00:00',  id_user int(11) NOT NULL default '0',  PRIMARY KEY  (ID),  KEY id_item (id_item),  KEY id_user (id_user),  KEY begin (begin),  KEY end (end));";
-
-	$db->query($query) or die("4202 ".$lang["update"][90].$db->error());
-}
 
 if(!FieldExists("glpi_tracking","device_type")) {
 	$query = "ALTER TABLE `glpi_tracking` ADD `device_type` INT DEFAULT '1' NOT NULL AFTER `assign` ;";
